+++ config.js
@@ -1,6 +1,7 @@
 export default {
   catalogUrl: null,
-  catalogTitle: "STAC Browser",
+  catalogTitle: null,
+  catalogTitleAfterImage: null,
   catalogImage: null,
   allowExternalAccess: true, // Must be true if catalogUrl is not given
   allowedDomains: [],
@@ -42,7 +43,8 @@ export default {
   pathPrefix: "/",
   historyMode: "history",
   cardViewMode: "cards",
-  cardViewSort: "asc",
+  defaultCollectionSort: "title",
+  defaultItemSort: null,
   showKeywordsInItemCards: false,
   showKeywordsInCatalogCards: false,
   showThumbnailsAsAssets: false,